Add timeout rule to WsnConnectorTest. Move OpeningHandshakeIT.java and WsnConnectorRule to a more logical package location (under specification/ws since this is a spec test).
