+++ transport/wsn/src/test/java/org/kaazing/gateway/transport/wsn/WsnConnectorTest.java
@@ -22,9 +22,11 @@
 package org.kaazing.gateway.transport.wsn;
 
 import static java.nio.ByteBuffer.wrap;
+import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.kaazing.gateway.util.Utils.asByteBuffer;
+import static org.kaazing.test.util.ITUtil.createTimeout;
 
 import java.net.URI;
 import java.nio.ByteBuffer;
@@ -34,7 +36,6 @@ import java.util.Properties;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
-import org.apache.log4j.PropertyConfigurator;
 import org.apache.mina.core.buffer.IoBuffer;
 import org.apache.mina.core.filterchain.IoFilterChain;
 import org.apache.mina.core.future.CloseFuture;
@@ -47,7 +48,6 @@ import org.apache.mina.core.session.IoSession;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
@@ -72,11 +72,10 @@ import org.kaazing.mina.core.buffer.IoBufferAllocatorEx;
 import org.kaazing.mina.core.buffer.IoBufferEx;
 import org.kaazing.mina.core.buffer.SimpleBufferAllocator;
 import org.kaazing.mina.core.session.IoSessionEx;
-import org.kaazing.test.util.MethodExecutionTrace;
 
 public class WsnConnectorTest {
     @Rule
-    public TestRule testExecutionTrace = new MethodExecutionTrace();
+    public TestRule cain = createTimeout(30, SECONDS);
 
     private SchedulerProvider schedulerProvider;
 
@@ -94,15 +93,6 @@ public class WsnConnectorTest {
     private HttpAcceptor httpAcceptor;
     private WsnAcceptor wsnAcceptor;
 
-    private static final boolean DEBUG = false;
-    @BeforeClass
-    public static void initLogging()
-            throws Exception {
-
-        if (DEBUG) {
-            PropertyConfigurator.configure("src/test/resources/log4j-trace.properties");
-        }
-    }
     @Before
     public void init() {
         schedulerProvider = new SchedulerProvider();
@@ -168,7 +158,7 @@ public class WsnConnectorTest {
         }
     }
 
-    @Test (timeout = 10000)
+    @Test
     public void shouldCloseWsnSessionWhenTransportClosesCleanlyButUnexpectedly() throws Exception {
 
         URI location = URI.create("wsn://localhost:8000/echo");
@@ -311,7 +301,7 @@ public class WsnConnectorTest {
 
     }
 
-    @Test (timeout = 30000)
+    @Test
     @Ignore("Failing on travis CI https://github.com/kaazing/gateway/issues/162")
     public void shouldNotHangOnToHttpConnectSessionsWhenEstablishingAndTearingDownWsnConnectorSessions() throws Exception {
